+++ pytrading212/position.py
@@ -1,6 +1,5 @@
 class Position:
-    def __init__(self, logo_url, instrument_code, value, quantity, total_return):
-        self.logo_url = logo_url
+    def __init__(self, instrument_code, value, quantity, total_return):
         self.instrument_code = instrument_code
         self.value = value
         self.quantity = quantity