+++ HISTORY.md
@@ -1,3 +1,7 @@
+### v0.1.4
+    * added order by value, Trading212 supports only value market order
+    * fix portfolio composition error
+    * removed logo_url from Position
 ### v0.1.3
     * minor fix
 ### v0.1.2